feat: StringCalulator클래스 main패키지로 이동
test 클래스에 있던 StringCalulator를 main클래스로 이동한 후, 테스트 확인 완료
